refactor : 사용하지 않는 DTO ì삭제
