+++ src/main/java/org/sopt/dto/comment/CommentEditRequest.java
@@ -1,4 +1,4 @@
-package org.sopt.dto;
+package org.sopt.dto.comment;
 
 public record CommentEditRequest(
 	String content
