+++ src/main/java/org/sopt/dto/post/PostRequest.java
@@ -1,8 +1,9 @@
-package org.sopt.dto;
+package org.sopt.dto.post;
 
 
 import org.sopt.domain.enums.Tag;
 
+// 게시글 생성 요청에 대한 DTO
 public record PostRequest(String title, String content, Tag tag){
 
 }
