+++ src/main/java/org/sopt/dto/post/PostResponse.java
@@ -1,14 +1,16 @@
-package org.sopt.dto;
+package org.sopt.dto.post;
 
 import java.util.List;
 
-import org.sopt.domain.Comment;
 import org.sopt.domain.Post;
+import org.sopt.dto.comment.CommentResponse;
+// 게시글 상세조회시 응답 DTO
 
 // 좋아요 수 추가됨
-public record PostCommentResponse(
+public record PostResponse(
 	Long id,
 	String title,
+	String content,
 	String userName,
 	List<CommentResponse> comments,
 	// 좋아요 추가
@@ -16,14 +18,15 @@ public record PostCommentResponse(
 
 ) {
 
-	public static PostCommentResponse from(Post post) {
+	public static PostResponse from(Post post) {
 		List<CommentResponse> commentDtos = post.getComments().stream()
 			.map(CommentResponse::from)
 			.toList();
 
-		return new PostCommentResponse(
+		return new PostResponse(
 			post.getId(),
 			post.getTitle(),
+			post.getContent(),
 			post.getUser().getName(),
 			commentDtos,
 			post.getLikeCount()
