+++ src/main/java/org/sopt/dto/post/PostSimpleResponse.java
@@ -1,9 +1,8 @@
-package org.sopt.dto;
+package org.sopt.dto.post;
 
 import org.sopt.domain.Post;
 
-import lombok.NoArgsConstructor;
-
+// 게시글에 대한 간단한 정보 표시할때 응답하는 DTO
 public record PostSimpleResponse(
     Long id,
     String title,
