+++ src/main/java/org/sopt/service/CommentLikeService.java
@@ -5,11 +5,8 @@ import static org.sopt.exception.ErrorCode.*;
 import org.sopt.domain.Comment;
 import org.sopt.domain.User;
 import org.sopt.domain.like.CommentLike;
-import org.sopt.domain.like.PostLike;
-import org.sopt.dto.CommentResponse;
-import org.sopt.dto.PostCommentResponse;
+import org.sopt.dto.comment.CommentResponse;
 import org.sopt.exception.CommentException;
-import org.sopt.exception.ErrorCode;
 import org.sopt.exception.UserException;
 import org.sopt.repository.CommentLikeRepository;
 import org.sopt.repository.CommentRepository;
