+++ src/main/java/org/sopt/service/UserService.java
@@ -2,7 +2,7 @@ package org.sopt.service;
 
 
 import org.sopt.domain.User;
-import org.sopt.dto.UserCreateRequest;
+import org.sopt.dto.user.UserCreateRequest;
 import org.sopt.repository.UserRepository;
 import org.springframework.stereotype.Service;
